Separated the `cuda_event` and the `event_pool` into their own header.
