+++ include/ghex/device/cuda/stream.hpp
@@ -21,36 +21,6 @@ namespace ghex
 {
 namespace device
 {
-struct cuda_event
-{
-    cudaEvent_t           m_event;
-    ghex::util::moved_bit m_moved;
-
-    cuda_event(){GHEX_CHECK_CUDA_RESULT(cudaEventCreateWithFlags(&m_event,
-        cudaEventDisableTiming))} cuda_event(const cuda_event&) = delete;
-    cuda_event& operator=(const cuda_event&) = delete;
-    cuda_event(cuda_event&& other) noexcept = default;
-    cuda_event& operator=(cuda_event&&) noexcept = default;
-
-    ~cuda_event()
-    {
-        if (!m_moved) { GHEX_CHECK_CUDA_RESULT_NO_THROW(cudaEventDestroy(m_event)) }
-    }
-
-    operator bool() const noexcept { return m_moved; }
-
-    cudaEvent_t& get() noexcept
-    {
-        assert(!m_moved);
-        return m_event;
-    }
-    const cudaEvent_t& get() const noexcept
-    {
-        assert(!m_moved);
-        return m_event;
-    }
-};
-
 /** @brief thin wrapper around a cuda stream */
 struct stream
 {
@@ -95,90 +65,5 @@ struct stream
         GHEX_CHECK_CUDA_RESULT(cudaStreamSynchronize(m_stream))
     }
 };
-
-/**
- * @brief	Pool of cuda events.
- *
- * Essentially a pool of events that can be used and reused one by one.
- * The main function is `get_event()` which returns an unused event.
- * To reuse an event the pool can either be rewinded, i.e. start again
- * with the first event, which requires that the user guarantees that
- * all events are no longer in use. The second way is to reset the pool
- * i.e. to destroy and recreate all events, which is much more expensive.
- *
- * Note that the pool is not thread safe.
- *
- * Todo:
- * - Maybe create a compile time size.
- * - Speed up `reset_pool()` by limiting recreation.
- */
-struct event_pool
-{
-  private: // members
-    std::vector<cuda_event> m_events;
-    std::size_t             m_next_event;
-    ghex::util::moved_bit   m_moved;
-
-  public: // constructors
-    event_pool(std::size_t expected_pool_size)
-    : m_events(expected_pool_size) // Initialize events now.
-    , m_next_event(0) {};
-
-    event_pool(const event_pool&) = delete;
-    event_pool& operator=(const event_pool&) = delete;
-    event_pool(event_pool&& other) noexcept = default;
-    event_pool& operator=(event_pool&&) noexcept = default;
-
-  public:
-    /** @brief	Get the next event of a pool.
-     *
-     * The function returns a new event that is not in use every time
-     * it is called. If the pool is exhausted new elements are created
-     * on demand.
-     */
-    cuda_event& get_event()
-    {
-        assert(!m_moved);
-        while (!(m_next_event < m_events.size())) { m_events.emplace_back(cuda_event()); };
-
-        const std::size_t event_to_use = m_next_event;
-        assert(!bool(m_events[event_to_use]));
-        m_next_event += 1;
-        return m_events[event_to_use];
-    };
-
-    /** @brief	Mark all events in the pool as unused.
-	 *
-	 * Essentially resets the internal counter of the pool, this means
-	 * that `get_event()` will return the very first event it returned
-	 * in the beginning. This allows reusing the event without destroying
-	 * and recreating them. It requires however, that a user can guarantee
-	 * that the events are no longer in use.
-	 */
-    void rewind()
-    {
-        if (m_moved) { throw std::runtime_error("ERROR: Can not reset a moved pool."); };
-        m_next_event = 0;
-    };
-
-    /** @brief	Clear the pool by recreating all events.
-	 *
-	 * The function will destroy and recreate all events in the pool.
-	 * This is more costly than to rewind the pool, but allows to reuse
-	 * the pool without having to ensure that the events are no longer
-	 * in active use.
-	 */
-    void clear()
-    {
-        if (m_moved) { throw std::runtime_error("ERROR: Can not reset a moved pool."); };
-
-        // NOTE: If an event is still enqueued somewhere, the CUDA runtime
-        //	will made sure that it is kept alive as long as it is still used.
-        m_events.clear();
-        m_next_event = 0;
-    };
-};
-
 } // namespace device
-
 } // namespace ghex
